requirements.txt
fastapi
uvicorn
aio_pika
firebase-admin
pydantic
python-multipart
redis
onnxruntime
pillow
numpy
pika
aioredis
python-dotenv